+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## v0.3beta2 2026-01-09
+- Turning on force HA update switch will override the in-built update HA button.
+- Added optional switch to force third party integration to update charging status in HA every 60 seconds instead of using the third party integration in-built update HA button, ie. possibly removes dependency on third party integration update rate.
 - Updated doc.
 
 ## v0.3beta1 2026-01-08